openai_realtime_api/requirements.txt
# OpenAI Realtime API Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
openai>=1.0.0
websocket-client>=1.6.0
pyaudio>=0.2.11

# Server dependencies (for ephemeral token server)
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0

# Example dependencies
keyboard>=0.13.5
asyncio-mqtt>=0.13.0

# Optional: WebRTC dependencies (for browser clients)
# aiortc>=1.5.0
# aiohttp>=3.8.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.5.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0 